Servicio registrar paciente completo

+++ src/main/java/com/sistemas/distribuidos/mediconnect/repositories/PacienteRepository.java
@@ -1,9 +1,10 @@
 package com.sistemas.distribuidos.mediconnect.repositories;
 
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.sistemas.distribuidos.mediconnect.models.PacienteModel;
 
 @Repository
-public interface PacienteRepository extends JpaRepository<PacienteModel, Integer>{
+public interface PacienteRepository extends JpaRepository<PacienteModel, Long> {
 }
